Update dtrt-indent-undo to process list of original variables
The `dtrt-indent-original-indent` variable is now list of lists of the
original values and localness of `indent-offset-variables`, including
`indent-tabs-mode`. Thus the latter is no longer a special case, and
`original-indent` being unset is no longer an edge case.

By making `dtrt-indent-original-indent` generic, the `dtrt-indent-undo`
function became much simpler, dealing with each element of the list as
originally local or not. Since this done in a map, the concatenation can
be done at the same time. Restoring `indent-tabs-mode` is no longer a
special case.