Explicitly check if key of generic list is bound
If not bound, it is not added to the list. This check is added because
`symbol-value` errors if the symbol's value is void (for instance, if
you don't have `evil-mode` installed). The macro `bound-and-true-p`
cannot be used because, as a macro, it would test the variable `mode`
instead of the value of the variable it references (`car x`).