Use case-sensitive search when applying language syntax table.
By default the emacs search functions are case-insensitive.  However
the dtrt-indent-language-syntax-table contains some keywords that
should be searched in case-sensitive manner.  This commit sets
case-fold-search to nil in dtrt-indent--for-each-indentation to make
those searches case-sensitive.

This affected the Erlang indent-guessing.  In Erlang, "fun" is a
keyword that introduces a lambda function that is terminated with
"end".  When dtrt-indent did a case-insensitive search, however, a
variable named "Fun" fooled it into thinking that a function was being
defined, and the rest of the file was not analyzed because there was
no matching "end".

This change will also affect the SGML CDATA pattern.