Fix skipping of escaped characters within strings.
Change the skip-regexp for strings in all languages from `"\\."` to
`"\\\\."`.  The intention of the skip-regexp in these cases is to
prevent dtrt-indent from incorrectly parsing an escaped quote
character inside a string as the end of the string.  To that end, it
needs to skip a literal `\` followed by the character being escaped,
so the regular expression is `\\.`.  Since this regular expression
then has to be represented as an emacs string, those backslashes have
to be escaped again, making `"\\\\."`.  To quote the elisp
documentation[1]:

    Note that `\' also has special meaning in the read syntax of Lisp
    strings (*note String Type::), and must be quoted with `\'.  For
    example, the regular expression that matches the `\' character is
    `\\'.  To write a Lisp string that contains the characters `\\',
    Lisp syntax requires you to quote each `\' with another `\'.
    Therefore, the read syntax for a regular expression matching `\'
    is `"\\\\"'.

[1] https://www.gnu.org/software/emacs/manual/html_node/elisp/Regexp-Special.html#index-g_t_0040samp_007b_005c_007d-in-regexp-3748