Allow offset to be forced to be a particular value
Add dtrt-indent-force-offset, which when set will allow the indent to
be set to a particular value. This value is used as if dtrt-indent
guessed it as the offset value.

This is intended to be used as a file-local or dir-local variable to
set the offset value when dtrt-indent comes up with the wrong value or
can't determine the value because only hard tabs are in use in the
file. Any related values are still set (e.g. evil-shift-width).

Logging of set value will show "[forced]" when this variable is used.