Fix testing of "tab-setting" when offset was rejected
Prior to this change it was not possible to use "expected-tab-setting"
in a test if a guessed offset was rejected, i.e. the threshold
`dtrt-indent-min-quality` was not reached.

The problem was that `dtrt-indent-functional-test` always checked if the
value of "expected-offset" is equal to the best guess even if
"expected-offset" was nil. There was no way to explictly check for a
rejected offset.

It is now possible to check if a guessed offset was rejected by setting
"expected-offset" to nil.