Allow change of tab-setting if offset was rejected
Since the very beginning, in d0eec66 (Initial git import, 2008-03-22), a
custom option exists to set the minimum amount of relevant lines.
According to the docstring this amount of lines is required for a guess
to be made. But here is the catch: It was never used during analysis. It
was only respected when performing a diagnosis. In this case though the
diagnosis lied since it was ignored in the actual analysis.

The commit beb7683 (Do not change anything when analyze is rejected,
2015-04-12) tried to ensure that empty buffers won't be changed by
dtrt-indent. This change introduced a regression: In case the guessed
offset was rejected, it stopped applying the tab-setting. But `rejected`
only concerns the guessed offset, it has nothing to do with the
tab-setting.

This commit fixes this regression by respecting the value of
`dtrt-indent-min-relevant-lines`.

In `dtrt-indent-functional-test` the default value of
`dtrt-indent-min-relevant-lines` (it is 2) is used because otherwise a
test would fail because it has only two relevant lines.